Add travis-ci build
This is the configuration for automated tests with travis-ci.org. It
currently builds Verilator and the RISC-V baremetal toolchain.

It then generates the simulation model, builds the Verilator
simulation and runs the assembler tests.

It also build the Linux toolchain, Linux and busybox, although they
are not used at the moment.