Fixes to set_env.sh
The primary issue was that `[ ! -d $UNDEFVAR ]` is always false, meaning that
$TOP wouldn't be set. I've adjusted the conditional. Additionally, I've
reformatted the opensocdebug related lines to use more conventional
indentation which matches the rest of the file. After discussion, we decided
no to quote variables at the moment in set_env because other scripts also need
updating, and it would be better if people encounter an error due to their
directory containing spaces early.